refactor: remove rename extension and add visual extension
Deleted the 'rename.js' extension and introduced a new 'visual.js' extension. The new extension enhances node creation by modifying node properties based on class names and titles, specifically for signature nodes.
